Cleanup newsletter logic.
