+++ public/scripts/newsletter.js
@@ -1,11 +1,11 @@
 // JS file for newsletter / files page
-//Set delete function
+// Set delete function
 // document.querySelectorAll(".deleteBtn").forEach((element, index) => {
 //     console.log(element);
 //     element.addEventListener('click', function () {
 //         $.ajax({
 //             type: 'DELETE',
-//             url: `/portal/newsletter/delete/${index}`,
+//             url: `/portal/newsletters/delete/${index}`,
 //             contentType: 'application/json'
 //         })
 //     });
