+++ views/partials/newsletter-upload-form.ejs
@@ -1,4 +1,4 @@
-<form action="/portal/newsletter/create" method="POST" class="p-3 m-3 form-card"
+<form action="/portal/newsletters/create" method="POST" class="p-3 m-3 form-card"
     style="background-color: #e9ecef" enctype="multipart/form-data">
 
     <div class="form-row text-center">
