+++ views/partials/portal-nav.ejs
@@ -16,7 +16,7 @@
 						<a class="nav-link" id="navlink" href="/portal/announcements">Announcements</a>
 					</li>
 					<li class="nav-item active">
-						<a class="nav-link" id="navlink" href="/portal/newsletter">Newsletters</a>
+						<a class="nav-link" id="navlink" href="/portal/newsletters">Newsletters</a>
 					</li>
 				<% } %>
