+++ views/portal/index.ejs
@@ -110,7 +110,7 @@
 
 						<h2 class="font-weight-bold font-175 mb-4">Newsletters</h2>
 
-						<a href="/portal/newsletter" class="portal-link d-block"
+						<a href="/portal/newsletters" class="portal-link d-block"
 							>Upload/Mange Newsletters <i class="fas fa-arrow-circle-right ml-1"></i
 						></a>
 					</div>
